+++ ext/process/ops/mod.rs
@@ -1,0 +1,1 @@
+// Copyright 2018-2025 the Deno authors. MIT license.
