+++ runtime/Cargo.toml
@@ -60,6 +60,7 @@ deno_kv.workspace = true
 deno_tls.workspace = true
 deno_url.workspace = true
 deno_web.workspace = true
+deno_process.workspace = true
 deno_webgpu.workspace = true
 deno_webidl.workspace = true
 deno_websocket.workspace = true
@@ -93,6 +94,7 @@ deno_node.workspace = true
 deno_os.workspace = true
 deno_path_util.workspace = true
 deno_permissions.workspace = true
+deno_process.workspace = true
 deno_resolver.workspace = true
 deno_telemetry.workspace = true
 deno_terminal.workspace = true
