+++ runtime/ops/mod.rs
@@ -4,7 +4,6 @@ pub mod bootstrap;
 pub mod fs_events;
 pub mod http;
 pub mod permissions;
-pub mod process;
 pub mod runtime;
 pub mod tty;
 pub mod web_worker;
