+++ runtime/ops/tty.rs
@@ -50,13 +50,12 @@ impl TtyModeStore {
   }
 }
 
+#[cfg(unix)]
+use deno_process::JsNixError;
 #[cfg(windows)]
 use winapi::shared::minwindef::DWORD;
 #[cfg(windows)]
 use winapi::um::wincon;
-
-#[cfg(unix)]
-use crate::ops::process::JsNixError;
 
 deno_core::extension!(
   deno_tty,
