+++ runtime/shared.rs
@@ -43,7 +43,6 @@ extension!(runtime,
     "10_permissions.js",
     "11_workers.js",
     "40_fs_events.js",
-    "40_process.js",
     "40_tty.js",
     "41_prompt.js",
     "90_deno_ns.js",
