+++ runtime/snapshot.rs
@@ -311,6 +311,7 @@ pub fn create_runtime_snapshot(
     deno_io::deno_io::init_ops_and_esm(Default::default()),
     deno_fs::deno_fs::init_ops_and_esm::<Permissions>(fs.clone()),
     deno_os::deno_os::init_ops_and_esm(Default::default()),
+    deno_process::deno_process::init_ops_and_esm(Default::default()),
     deno_node::deno_node::init_ops_and_esm::<
       Permissions,
       DenoInNpmPackageChecker,
@@ -325,7 +326,6 @@ pub fn create_runtime_snapshot(
     ),
     ops::fs_events::deno_fs_events::init_ops(),
     ops::permissions::deno_permissions::init_ops(),
-    ops::process::deno_process::init_ops(None),
     ops::tty::deno_tty::init_ops(),
     ops::http::deno_http_runtime::init_ops(),
     ops::bootstrap::deno_bootstrap::init_ops(Some(snapshot_options)),
