+++ tools/core_import_map.json
@@ -241,7 +241,7 @@
     "ext:runtime/11_workers.js": "../runtime/js/11_workers.js",
     "ext:deno_os/30_os.js": "../ext/os/30_os.js",
     "ext:runtime/40_fs_events.js": "../runtime/js/40_fs_events.js",
-    "ext:runtime/40_process.js": "../runtime/js/40_process.js",
+    "ext:deno_process/40_process.js": "../ext/process/40_process.js",
     "ext:deno_os/40_signals.js": "../ext/os/40_signals.js",
     "ext:runtime/40_tty.js": "../runtime/js/40_tty.js",
     "ext:runtime/41_prompt.js": "../runtime/js/41_prompt.js",
